Updated the readme
# Hypixel-Autoguild
This is the code for the bot that I used to run the Lostshard Hypixel guild. 
Hypixel has limited guild invites to 10 every 5 minutes, so this code no longer works.
However, I have provided it for reference.

This is definently spaghetti code. When I wanted to add ipmlement new functoinality, I woudld just tack it onto the script.
It didn't seem worthwhile to rewrite the entire project.